Created On Duty Report instead of using Timesheet::findForQuery. (#1878)
- Indicate when an entry exceeds 1.5x an associated shift sign up or 6 hours (i.e., 1.5*6 = 9 hours).
- Return the associated shift if any

+++ app/Models/TimesheetMissing.php
@@ -257,7 +257,7 @@ class TimesheetMissing extends ApiModel
         $on_duty = $this->getOriginal('on_duty');
         $off_duty = $this->getOriginal('off_duty');
 
-        return Carbon::parse($on_duty)->diffInSeconds(Carbon::parse($off_duty));
+        return (int) Carbon::parse($on_duty)->diffInSeconds(Carbon::parse($off_duty));
     }
 
     /**
